requirements.txt
streamlit
jupyter-dash
google-generativeai